# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Server API",
        "description": "API сайта",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://siveno.shop"
        }
    ],
    "paths": {
...
        "/api/colors": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Получение списка всех цветов",
                "operationId": "9af5ded8d7143a0f97e53725eda97464",
                "responses": {
                    "200": {
                        "description": "Список цветов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Красный"
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "#FF0000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Colors"
                ],
                "summary": "Создание нового цвета",
                "operationId": "e4d3770c69f36a71c5936b3b35655342",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Синий"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "#0000FF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Цвет создан"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/colors/{id}": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Получение информации о конкретном цвете",
                "operationId": "c0e5a39d91e5093e2e0f77823615db5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цвета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о цвете"
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Colors"
                ],
                "summary": "Обновление информации о цвете",
                "operationId": "717be7234f38d1fb4ff4678ba4f59581",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цвета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Зеленый"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "#00FF00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Цвет обновлен"
                    },
                    "404": {
                        "description": "Цвет не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Colors"
                ],
                "summary": "Удаление цвета",
                "operationId": "1667d6964960e3a8158e80e50c69f19c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цвета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Цвет удален"
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            }
        },
        "/api/colors/{id}/products": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Получение всех продуктов, связанных с цветом",
                "operationId": "8e4e8ba54944cd1d7fd8dc9c129b1971",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цвета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список продуктов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Футболка"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 19.989999999999998436805981327779591083526611328125
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}